hangman_prj/hangman/api/serializers.py
from rest_framework import serializers
from hangman.models import Game, Leaderboard


class GameSerializer(serializers.ModelSerializer):
    class Meta:
        model = Game
        fields = '__all__'


class LeaderboardSerializer(serializers.ModelSerializer):
    class Meta:
        model = Leaderboard
        fields = '__all__'